Update index.html
To add an element that counts the number of site visits, you can use local storage in JavaScript to keep track of the number of visits to the site. This will only count visits on the same browser, as local storage is specific to the browser and device.

Here's how you can implement this:

Add an HTML element where the visit count will be displayed.